# chore: make materialized cte scan_progress accurate