--- a/src/query/pipeline/sources/src/async_source.rs
+++ b/src/query/pipeline/sources/src/async_source.rs
@@ -116,7 +116,9 @@ impl<T: 'static + AsyncSource> Processor for AsyncSourcer<T> {
         match self.inner.generate().await? {
             None => self.is_finish = true,
             Some(data_block) => {
-                if !data_block.is_empty() {
+                // Don't need to record the scan progress of `MaterializedCteSource`
+                // Because it reads data from memory.
+                if !data_block.is_empty() && self.name() != "MaterializedCteSource" {
                     let progress_values = ProgressValues {
                         rows: data_block.num_rows(),
                         bytes: data_block.memory_size(),